revert prev. commit and extract message state to a new file
